refactor(front): extract useIsMorphRelationReadOnlyFromRelatedRecordPerspective hook (#24231)
Extracts a shared hook to encapsulate morph relation read-only rule logic for both FieldWidgetMorphRelationCard and RecordDetailMorphRelationSectionDropdown.
